refactor: refactor new devices and docs
