+++ docs/usage.md
@@ -113,19 +113,19 @@ async def control_breeze(device_ip, device_id, remote_manager, remote_id) :
         remote = remote_manager.get_remote(remote_id)
         # prepare a control command that turns on the Breeze
         # set to 24 degree (Celsius) cooling with vertical swing
-        # and keep the current Fan Level (3)
-        command: SwitcherBreezeCommand = remote.get_command(
+        # send command to the device (3)
+        await api.control_breeze_device(
+            remote,
             DeviceState.ON,
             ThermostatMode.COOL,
             24,
-            resp.fan_level,
+            ThermostatFanLevel.MEDIUM,
             ThermostatSwing.ON,
-            response.state
         )
-        # send command to the device (4)
-        await api.control_breeze_device(command)
+        # if you're writing stateful software,
+        # you can include the previous state for toggle devices
 
-# create the remote manager outside the context for re-using (5)
+# create the remote manager outside the context for re-using (4)
 remote_manager = SwitcherBreezeRemoteManager()
 asyncio.get_event_loop().run_until_complete(
     control_breeze("111.222.11.22", "ab1c2d", remote_manager, "DLK65863")
@@ -134,9 +134,8 @@ asyncio.get_event_loop().run_until_complete(
 
 1. [SwitcherThermostatStateResponse](./codedocs.md#src.aioswitcher.api.messages.SwitcherThermostatStateResponse)
 2. [SwitcherBreezeRemote](./codedocs.md#src.aioswitcher.api.messages.SwitcherBreezeRemote)
-3. [SwitcherBreezeCommand](./codedocs.md#src.aioswitcher.api.SwitcherBreezeCommand)
-4. [SwitcherBaseResponse](./codedocs.md#src.aioswitcher.api.messages.SwitcherBaseResponse)
-5. [SwitcherBreezeRemoteManager](./codedocs.md#src.aioswitcher.api.SwitcherBreezeRemoteManager)
+3. [SwitcherBaseResponse](./codedocs.md#src.aioswitcher.api.messages.SwitcherBaseResponse)
+4. [SwitcherBreezeRemoteManager](./codedocs.md#src.aioswitcher.api.SwitcherBreezeRemoteManager)
 
 !!! info
     You can find the supported device types stated in [this enum](./codedocs.md#src.aioswitcher.device.DeviceType) members.
