+++ tests/testresources/dummy_responses/control_breeze_swing_response.txt
@@ -1,0 +1,1 @@
+fef0300004000102000000000000020000000000000000008d6a966200000000000000000000f0fe01000000f5c7f750
